sorting widgets on login_screen
